backend/main.py
from fastapi import FastAPI
from pydantic import BaseModel
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv
import os

load_dotenv()
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
app = FastAPI()
llm = ChatOpenAI(model_name="gpt-3.5-turbo", openai_api_key=os.getenv("OPENAI_API_KEY"))

class QueryRequest(BaseModel):
  question: str

@app.post("/ask")
def ask_question(request: QueryRequest):
  response = llm(request.question)
  return {"answer": response}
